refactor: 관리자 정보 카드 컴포넌트 분리
