+++ src/pages/Manager/ManagedClubDetail/index.tsx
@@ -3,7 +3,7 @@ import BoyIcon from '@/assets/image/boy.png';
 import ChatIcon from '@/assets/image/chat.png';
 import FloderIcon from '@/assets/image/folder.png';
 import ChevronRightDarkIcon from '@/assets/svg/Chevron-left-dark.svg';
-import UserInfoCard from '@/pages/User/MyPage/components/UserInfoCard';
+import ManagerInfoCard from '@/pages/User/MyPage/components/ManagerInfoCard';
 
 const menuItems = [
   { to: 'recruitment', icon: FloderIcon, size: 24, label: '모집 공고 및 지원서 관리' },
@@ -14,7 +14,7 @@ const menuItems = [
 function ManagedClubDetail() {
   return (
     <div className="flex flex-col gap-9 px-[19px] py-[17px]">
-      <UserInfoCard type="detail" />
+      <ManagerInfoCard type="detail" />
       <div className="border-background flex flex-col gap-5 rounded-2xl border bg-white px-4 py-3">
         {menuItems.map(({ to, icon, size, label }) => (
           <Link key={to} to={to} className="flex items-center justify-between">
